fix: avoid concurrency during initialization
using a filelock to ensure that initialization of the same version happens only once.

also improved debug logging a bit

+++ cliv2/internal/cliv2/cliv2.go
@@ -14,10 +14,12 @@ import (
 	"regexp"
 	"strings"
 
+	"github.com/gofrs/flock"
 	"github.com/snyk/cli/cliv2/internal/constants"
 	"github.com/snyk/cli/cliv2/internal/embedded"
 	"github.com/snyk/cli/cliv2/internal/embedded/cliv1"
 	"github.com/snyk/cli/cliv2/internal/proxy"
+	local_utils "github.com/snyk/cli/cliv2/internal/utils"
 	"github.com/snyk/go-application-framework/pkg/utils"
 )
 
@@ -62,15 +64,54 @@ func NewCLIv2(cacheDirectory string, debugLogger *log.Logger) (*CLI, error) {
 		stderr:           os.Stderr,
 	}
 
-	cli.ClearCache()
+	return &cli, nil
+}
 
-	err = cli.ExtractV1Binary()
-	if err != nil {
-		fmt.Println(err)
-		return nil, err
+func (c *CLI) Init() (err error) {
+	c.DebugLogger.Println("Init start")
+
+	// ensure the specified base cache directory exists, this needs to be done even before acquiring the lock
+	if _, err = os.Stat(c.CacheDirectory); os.IsNotExist(err) {
+		err = os.Mkdir(c.CacheDirectory, local_utils.CACHEDIR_PERMISSION)
+		if err != nil {
+			return err
+		}
 	}
 
-	return &cli, nil
+	// use filelock to synchronize parallel executed processes
+	lockFileName := path.Join(c.CacheDirectory, GetFullVersion()+".lock")
+	fileLock := flock.New(lockFileName)
+	err = fileLock.Lock()
+	if err != nil {
+		return err
+	}
+
+	unlock := func() {
+		fileLock.Unlock()
+		os.Remove(lockFileName)
+	}
+	defer unlock()
+
+	c.DebugLogger.Printf("Init-Lock acquired: %v (%s)\n", fileLock.Locked(), lockFileName)
+
+	// create required cache and temp directories
+	err = local_utils.CreateAllDirectories(c.CacheDirectory, GetFullVersion())
+	if err != nil {
+		return err
+	}
+
+	// cleanup cache a bit
+	c.ClearCache()
+
+	// extract cliv1
+	err = c.ExtractV1Binary()
+	if err != nil {
+		return err
+	}
+
+	c.DebugLogger.Println("Init end")
+
+	return err
 }
 
 func (c *CLI) ClearCache() error {
@@ -83,16 +124,18 @@ func (c *CLI) ClearCache() error {
 	// Get current version binary's path
 	v1BinaryPath := path.Dir(c.v1BinaryLocation)
 	var maxVersionToDelete = 5
-	for i, file := range fileInfo {
+	var deleteCount = 0
+	for _, file := range fileInfo {
 		currentPath := path.Join(c.CacheDirectory, file.Name())
-		if currentPath != v1BinaryPath {
+		if currentPath != v1BinaryPath && !strings.Contains(currentPath, ".lock") {
+			deleteCount++
 			err = os.RemoveAll(currentPath)
 			if err != nil {
 				c.DebugLogger.Println("Error deleting an old version directory: ", currentPath)
 			}
 		}
 		// Stop the loop after 5 deletions to not create too much overhead
-		if i == maxVersionToDelete {
+		if deleteCount == maxVersionToDelete {
 			break
 		}
 	}
@@ -105,7 +148,7 @@ func (c *CLI) ExtractV1Binary() error {
 
 	isValid, err := embedded.ValidateFile(c.v1BinaryLocation, cliV1ExpectedSHA256, c.DebugLogger)
 	if err != nil || !isValid {
-		c.DebugLogger.Println("cliv1 is not valid, start extracting ", c.v1BinaryLocation)
+		c.DebugLogger.Println("Extract cliv1 to", c.v1BinaryLocation)
 
 		err = cliv1.ExtractTo(c.v1BinaryLocation)
 		if err != nil {
@@ -118,13 +161,13 @@ func (c *CLI) ExtractV1Binary() error {
 		}
 
 		if isValid {
-			c.DebugLogger.Println("cliv1 is valid after extracting", c.v1BinaryLocation)
+			c.DebugLogger.Println("Extracted cliv1 successfully")
 		} else {
-			fmt.Println("cliv1 is not valid after sha256 check")
+			c.DebugLogger.Println("Extracted cliv1 is not valid")
 			return err
 		}
 	} else {
-		c.DebugLogger.Println("cliv1 already exists and is valid at", c.v1BinaryLocation)
+		c.DebugLogger.Println("Extraction not required")
 	}
 
 	return nil
